Move code to more suited place.
